# Speed up Dataset._construct_dataarray