@dcherian 2021-01-05
Merging. This logic could be updated if we convert `_coord_names` to a list or OrderedSet.